Implement a filter for violations
In order to enable the user to specify which dependencies should not break the build (i.e. are not forbidden; see issue #1) the violations returned by JDeps must be filtered. The new class 'ViolationFilter' requires an instance that decides for individual dependencies whether they are forbidden (called 'DependencyFilter' for now) and uses it to create filtered violations.